improve item response hande
