+++ src/main/java/itu/eval_2/newapp/models/action/FrappeModel.java
@@ -1,7 +1,31 @@
 package itu.eval_2.newapp.models.action;
 
-public interface FrappeModel {
+import java.time.LocalDateTime;
 
-    public void cotnrole();
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import itu.eval_2.newapp.annotations.date.ErpNextDateTime;
+import lombok.Data;
+
+@Data
+public abstract class FrappeModel {
+    // Document metadata
+    private String name;
+    private String owner;
     
+    @ErpNextDateTime
+    private LocalDateTime creation;
+    
+    @ErpNextDateTime
+    private LocalDateTime modified;
+    
+    @JsonProperty("modified_by")
+    private String modifiedBy;
+    
+    // Document status
+    private int docstatus;
+    private int idx;
+    private String title;
+    private String status;
+    abstract public void cotnrole();    
 }
