+++ src/main/java/itu/eval_2/newapp/services/frappe/quotation/ErpNextQuotationServiceImpl.java
@@ -5,8 +5,6 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import itu.eval_2.newapp.config.ApiConfig;
-import itu.eval_2.newapp.config.FrappApiFilter;
-import itu.eval_2.newapp.config.FrappeApiFilterList;
 import itu.eval_2.newapp.exceptions.ERPNextIntegrationException;
 import itu.eval_2.newapp.models.api.requests.UpdateQuotationRequest;
 import itu.eval_2.newapp.models.api.responses.SingleSupplierQuotationResponse;
